[ADL-95] Rename some commands
add-{wallet,account,address} commands have been renamed to
new-{wallet,account,address}.
Their semantics is to generate an entity and add it to the storage.
The "restore" command also adds a wallet, but not a new one.
The primary reason is that we want to have two different functions in
code with different semantics.