[AD-239] Move ariadne-specific tests from knit to ariadne
Otherwise knit tests don't compile, because stack misbehaves when a
a test suite from package A depends on library A and library B which
also depends on library A.
Knit tests shouldn't depend on ariadne anyway.
Proper soultion is to keep tests for Core in knit and move the rest
to ariadne, but it will require a bit more refactoring which is out of
scope of AD-239.