[AD-234] Generate a new address as change
Previously we were reusing one of our addresses as change address,
which is considered bad practice.
Proper approach is to generate a new change address for each transaction.
This logic is handled by the `MonadAddresses` class.
For simplicity an `IO` action which generates a new address is used
as `AddrData`. This action is `walletNewAddress` from `WalletFace` with
proper arguments. `walletNewAddress` was modified to return the address
it generates.