[AD-328] Leave only essential stuff in CardanoConfig
CardanoConfig is not a newtype over CommonNodeArgs anymore, because
we don't want to have some things from CommonNodeArgs configurable
(and because it will be needed for AD-205 later anyway).
I also made it almost flat (e. g. there is nothing analogous to CommonArgs),
because smaller types are only used as part of CommonNodeArgs, not by
themselves (in the code we are interested in).