[AD-328] Disable unused-top-binds warning only in specific modules
This warning is a good warning, it makes it easy to delete unused stuff,
which is a good thing to do. Without this warning it becomes hard to track
which code is actually used. And currently there are not many modules where
suppression of this warning is actually needed, so the easiest thing to do
is to suppress it only in those modules.
Note: reasonable use case to suppress this warning is when you generate lenses
for some type and need only a few of them.