[AD-273] Now the whole subtree is deleted when someone deletes root of that subtree.
For example, in case of wallet's deletion, accounts and addresses will be also deleted.
Deleteing of the whole subtree happens atomically.
I use "deleteIxAll" function to delete unused addresses and accouns directly.
But this function des nothing if the key is not presented in IxSet (invalid HdRootId/HdAccountId),
that is why i check it in the beginning and throw exception in case of invalid id.