[AD-350] Make two defaults configs identical
Previously there was a small difference between using
`config/ariadne-config.dhall` and using hardcoded config (which is used
when config file can't be found), because relative paths in hardcoded
config weren't treated as relative to 'configPath', while paths in
config file were.
Having two different default configs seems bad and confusing, so this
difference was eliminated.
The only relative path in the default config is path to
`cardano-configuration.yaml`. In default case the only difference for a user
is that this file (and `mainnet-genesis.json`) will be dumped into
`~/.config/ariadne/` instead of current directory, which is better, I think.