[AD-382] Add tool for viewing acid-state contents
This commit adds utility that displays acid state DB with hdwallets.
It will be useful for debugging.
Also it can show all previous states of DB (up to the last checpoint).
See readme with examples of using it.

Commit was updateв due to PR reviews
  (small changes,fixes in readme, etc).
New update due to update of corresponding acid-state commit.
Upd: changed default path to "wallet-db", instead of ".wallet-db"
     And auto-detect home path.