[AD-502] Fix the bug with `getstate` function.
Problem: In PR #350 were added the following bugs:
  1. Ariadne failed to be started with an empty database.
  2. `openLocalState` function loads not the current state of DB, but the previous one.

The cause of this bugs is `getState` function,
which returns not the given state but the previous.

Solution: Bugfix was added in the new PR to the acid-state repository,
and new commit replaced the old one in stack.yaml.
Also, there was added 2 tests to the `backend` test group:
test of opening an empty database and test of loading the last state of DB
after closing it.