[AD-520] Update version of the application
Problem: version 0.3.0.0 was released recently, but the version in hpack
file wasn't updated. Our simple update system treats this value as the current
version of the application and compares it to the latest released one. So
it's important to maintain it properly.
Solution: update 'version' in hpack/definitions.yaml and CHANGELOG accordingly.