[AD-519] Update dhall version
Problem: we want to use `co-log` for logging which requires a newer
`contravariant` version. The version of `dhall` which we are currently using
is not compatible with this `contravariant` version. Hence it's necessary to
update `dhall`.
There was a breaking change in 1.14.0 which switched grammar of `Natural` and
`Integer`.
See the CHANGELOG: https://github.com/dhall-lang/dhall-haskell/blob/83c4a32e99ae0bea27c1fb548ce58c13296eff5b/dhall/CHANGELOG.md
Solution: update `dhall` version to `1.15.1`, which is the first one
compatible with `contravariant-1.5.0`.
Also update static configs in this repository according to the breaking
changes.