[AD-445] Change quit command handling
Problem: when a quit command is entered in REPL, it will call
`putUiEvent` from the UI thread. `putUiEvent` may block if the
queue is full. It will lead to deadlock, because UI thread is the only
thread that reads from this queue. Apart from that such indirect approach
seems more complicated.
Solution: add one more `WidgetEventResult` and return it from
`handleReplWidgetKey` when quit command is handled. It allowed to get rid
of `UiCommandQuit` and of `UiFace` in REPL widget.