[AD-445] Do not send command to REPL from UI thread
Problem: `langPutUiCommand` is executed from the UI thread and it calls
`putUiEvent` which can block and only the UI thread can unblock it.
So deadlock is possible.
Solution: call `putUiEvent` from Knit backend, inside action passed to
`spawnTask`. That action is executed from non-UI thread, so it can't deadlock.
One alternative would be to introduce a way to send an event from widgets to
REPL. However, that seems to be too hard.
Another alternative is to fork a thread specifically for `putUiEvent`, but
this event should be processed before the command is executed, because
otherwise `updateCommandResult` will ignore all events related to this command.
This problem was present before, because `putUiEvent` was called after
`putCommand`.
This approach is quite simple and doesn't require additional thread management.